requirements.txt
Django==2.1.7
django-birthday==0.1.3
django-ckeditor==5.6.1
django-phonenumber-field==2.3.0
django-widget-tweaks==1.4.3
djangorestframework==3.9.2
djangorestframework_simplejwt==4.3.0
phonenumbers==8.10.8
Pillow==5.4.1
PyJWT==1.7.1
python-decouple==3.1
social-auth-app-django==3.1.0